+++ lib/ccxt_extract/oxc_extractor.ex
@@ -78,11 +78,7 @@ defmodule CcxtExtract.OXCExtractor do
         {exchanges, skipped, errors} =
           files
           |> Enum.map(&parse_file/1)
-          |> Enum.reduce({[], [], []}, fn
-            {:ok, exchange}, {ok, skip, err} -> {[exchange | ok], skip, err}
-            {:skip, file}, {ok, skip, err} -> {ok, [file | skip], err}
-            {:error, file, reason}, {ok, skip, err} -> {ok, skip, [{file, reason} | err]}
-          end)
+          |> CcxtExtract.OXCBatch.reduce_results()
 
         for {file, reason} <- errors do
           Logger.warning("Failed to parse #{file}: #{inspect(reason)}")
@@ -101,21 +97,7 @@ defmodule CcxtExtract.OXCExtractor do
       """
       @spec parse_file(String.t()) ::
               {:ok, map()} | {:skip, String.t()} | {:error, String.t(), term()}
-      def parse_file(path) do
-        source = File.read!(path)
-        filename = Path.basename(path)
-
-        case OXC.parse(source, filename) do
-          {:ok, ast} ->
-            case extract_from_ast(ast, filename) do
-              nil -> {:skip, filename}
-              exchange -> {:ok, exchange}
-            end
-
-          {:error, reason} ->
-            {:error, filename, reason}
-        end
-      end
+      def parse_file(path), do: CcxtExtract.OXCBatch.parse_file(path, &extract_from_ast/2)
 
       @doc """
       Write extracted data to the discovery JSON file.